+++ lib/make_taggable.rb
@@ -47,7 +47,6 @@ module MakeTaggable
     autoload :Core
     autoload :Ownership
     autoload :Related
-    autoload :TagListType
   end
 
   autoload :Utils